Add files via upload
Description: This project focuses on using machine learning techniques to classify objects as either rocks or mines based on sonar signals. The dataset used for this project consists of sonar returns mines or rocks, and the goal is to develop a predictive model that can accurately distinguish between the two. Here we are using logistic regression model